Fix double value bug

+++ src/com/company/Student.java
@@ -15,7 +15,7 @@ public class Student {
     }
 
     public double calculateAverage () {
-        return (this.gradeEnglish + this.gradeMath + this.gradeSport) / 3;
+        return (double)(this.gradeEnglish + this.gradeMath + this.gradeSport) / 3;
     }
 
     public String getName() {
